% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaols.R
\name{avaols}
\alias{avaols}
\title{All-vs-All Ordinary Least Squares (AVAOLS) Regression}
\usage{
avaols(X, intercept = TRUE)
}
\arguments{
\item{X}{a matrix or data frame of predictors}

\item{intercept}{a logical value indicating whether or not to include an intercept in the model. Default is TRUE.}
}
\value{
an object of class "avaols" containing the following components:
\item{X}{a copy of the input data frame}
\item{B}{a matrix of coefficients}
\item{B0}{a matrix of intercepts}
}
\description{
This function performs an All-vs-All Ordinary Least Squares (AVAOLS) Regression
using the input matrix X.
The model is fitted in one step using the QR decomposition, instead of naively fitting model for each columns separately.
}
\examples{
X <- data.frame(matrix(rnorm(100), ncol = 5))
avaols(X)
}
